Moved setUserDetailsWithRole for now. (to refactor later)
